+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* Allow specifying custom resource attributes on `GrafanaOpenTelemetrySettings`.
 * Improve accuracy of resource attributes `telemetry.distro.name` and `telemetry.distro.version`.
 
 ### Bug fixes